Refactor scripts directory (#306)
* Refactor scripts directory
* Execute test_assets via subprocess
