+++ scripts/release/requirements.txt
@@ -1,0 +1,2 @@
+#azureml-assets
+azure-ai-ml>=1.2.0
