--- scripts/requirements.txt
@@ -1,3 +1,0 @@
-GitPython==3.1.27
-pyyaml==5.4
-requests==2.27.1
