+++ scripts/test/requirements.txt
@@ -1,0 +1,2 @@
+#azureml-assets
+pytest>=7.1.1
