+++ scripts/validation/requirements.txt
@@ -1,0 +1,2 @@
+flake8>=4.0.1
+pydocstyle>=6.1.1
